% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{age_at_death}
\alias{age_at_death}
\title{Age at death}
\usage{
age_at_death(data, instance_num, death_date_field = f.40000.0.0.Date_of_death)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{instance_num}{An integer or string specifying the instance number, or the <\code{\link[=dplyr_data_masking]{data-masking}}> name of the column containing instance numbers to use in the current function.

The instance number is also known as the assessment visit number, of which there are 4 (labeled 0...3) in the UK BioBank study. This column does not exist in a standard distribution of the UKB data set, and should be defined and joined to the data set before performing analyses with this function.}

\item{death_date_field}{e.g. f.40000.0.0.Date_of_death}
}
\description{
Age at death
}
