% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age_at_instance}
\alias{age_at_instance}
\alias{age_at_instance_computed}
\title{Age at instance}
\usage{
age_at_instance(
  data,
  instance_num,
  age_at_instance_field = f.21003.0.0.Age_when_attended_assessment_centre
)

age_at_instance_computed(
  data,
  instance_num,
  date_of_instance_field = f.53.0.0.Date_of_attending_assessment_centre,
  year_of_birth_field = f.34.0.0.Year_of_birth,
  month_of_birth_field = f.52.0.0.Month_of_birth
)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{instance_num}{An integer or string specifying the instance number, or the <\code{\link[=dplyr_data_masking]{data-masking}}> name of the column containing instance numbers to use in the current function.

The instance number is also known as the assessment visit number, of which there are 4 (labeled 0...3) in the UK BioBank study. This column does not exist in a standard distribution of the UKB data set, and should be defined and joined to the data set before performing analyses with this function.}

\item{age_at_instance_field}{Template field (using <\code{\link[=dplyr_data_masking]{data-masking}}> rules) for the age at each instance.}

\item{date_of_instance_field}{Template field (using <\code{\link[=dplyr_data_masking]{data-masking}}> rules) for the date associated with each instance.}

\item{year_of_birth_field}{<\code{\link[=dplyr_data_masking]{data-masking}}> Name of column containing birth year}

\item{month_of_birth_field}{<\code{\link[=dplyr_data_masking]{data-masking}}> Name of column containing birth month}
}
\description{
Age, in years, at the time of the specific instance.
}
\details{
\code{age_at_instance} uses the coded fields that contain the age (rounded to the nearest year) at each instance visit.
\code{age_at_instance_computed} calculates a slightly more precise age based on the partipants year and month of birth.
}
