% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoses.R
\name{age_of_diagnosis_placeholder}
\alias{age_of_diagnosis_placeholder}
\alias{age_of_first_diagnosis}
\alias{age_of_first_mi}
\alias{age_of_first_stroke}
\alias{age_of_first_chf}
\alias{age_of_first_copd}
\title{Age of first diagnosis for selected diseases}
\usage{
age_of_first_diagnosis(
  data,
  diagnosis_date_col,
  year_of_birth_col = f.34.0.0.Year_of_birth,
  month_of_birth_col = f.52.0.0.Month_of_birth
)

age_of_first_mi(
  data,
  diagnosis_date_col = f.42000.0.0.Date_of_myocardial_infarction,
  year_of_birth_col = f.34.0.0.Year_of_birth,
  month_of_birth_col = f.52.0.0.Month_of_birth
)

age_of_first_stroke(
  data,
  diagnosis_date_col = f.42006.0.0.Date_of_stroke,
  year_of_birth_col = f.34.0.0.Year_of_birth,
  month_of_birth_col = f.52.0.0.Month_of_birth
)

age_of_first_chf(
  data,
  diagnosis_date_col = f.131354.0.0,
  year_of_birth_col = f.34.0.0.Year_of_birth,
  month_of_birth_col = f.52.0.0.Month_of_birth
)

age_of_first_copd(
  data,
  diagnosis_date_col = f.42016.0.0.Date_chronic_obstructive_pulmonary_disease_report,
  year_of_birth_col = f.34.0.0.Year_of_birth,
  month_of_birth_col = f.52.0.0.Month_of_birth
)
}
\arguments{
\item{data}{The primarydata frame.

Thisdata frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{year_of_birth_col}{Template column name for year of birth.\cr Default = \code{f.34.0.0.Year_of_birth}.}

\item{month_of_birth_col}{Template column name for month of birth.\cr Default = \code{f.52.0.0.Month_of_birth}.}
}
\description{
These convenience functions use Category 42 ("Algorithmically
define outcomes") to calculate the age at which a particular disease
condition was diagnosed.
}
\section{Instancing}{
 The UK Biobank records visits as separate "instances." As of this writing, there are 4 instances labeled 0 through 3. At each instance, various information can be recorded or re-recorded. For example, blood pressure is typically recorded at most in-person evaluations. Therefore, there may be 4 separate columns for blood pressure recordings (actually, there could be more because the blood pressure may be recorded multiple times at each instance). Almost all of the functions in this package will utilize instance numbers to specify from which time pointsdata should be retrieved. For example, we may want to know the state of ICD10 diagnoses \emph{before} instance 2. In this case, we would specify \code{up_to_instance = 1} (search up to instance 1, inclusive) in functions that take this as an argument.

Arguments like \code{up_to_instance} and \code{after_instance} can take a constant instance number as their value. But they can also take the name of a column that contains an instance number, so that different instance limits can be used for each participant. For example, some participants undergo MRI at instance 2, and others at instance 3. If we want to know the state of a diagnosis up to and including the time of MRI, we would want to assign \code{up_to_instance} to the name of the column that specifies which instance the MRI occurred at. This column typically has to be generated by the user and attached to thedata frame beforehand.
}

\section{Column names}{
 Most functions in this package will take column names (\verb{*_col}) as optional arguments (otherwise a default column names are used) which are then used as templates to find all other columns with the same field number, but different instance and array numbers. These functions will automatically find all matching instances (and arrays within each instance) within the specified parameters. Internally, a set of \verb{expand_instance_*()} helper functions, which themselves rely on the \code{column_expansion_helper()} function, perform the work of searching for matching columns.
}

