% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{any_by_row}
\alias{any_by_row}
\alias{fn_by_row}
\alias{reduce_by_row}
\title{Apply functions across rows}
\usage{
any_by_row(data)

fn_by_row(data, fn)

reduce_by_row(data, fn)
}
\arguments{
\item{data}{input data}

\item{fn}{function to apply}
}
\description{
\itemize{
\item \code{any_by_row}:  For each row, are any of the columns true (see \code{\link[=any]{any()}})
\item \code{fn_by_row}:  Apply function across each entire row (function takes whole row as argument)
\item \code{reduce_by_row}:  Apply \code{\link[=Reduce]{Reduce()}} across each row of matrix or dataframe
}
}
\keyword{internal}
