% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{biomarker_lookup}
\alias{biomarker_lookup}
\alias{biomarker_A1c_percent}
\alias{biomarker_LDL_mgdL}
\alias{biomarker_SCr_mgdL}
\title{Biomarker lookups}
\usage{
biomarker_lookup(
  data,
  biomarker_field,
  combine_instances = c("last", "first", "min", "max", "mean"),
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST
)

biomarker_A1c_percent(
  data,
  biomarker_field = f.30750.0.0.Glycated_haemoglobin_HbA1c,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  combine_instances = c("last", "first", "min", "max", "mean")
)

biomarker_LDL_mgdL(
  data,
  biomarker_field = f.30780.0.0.LDL_direct,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  combine_instances = c("last", "first", "min", "max", "mean")
)

biomarker_SCr_mgdL(
  data,
  biomarker_field = f.30700.0.0.Creatinine,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  combine_instances = c("last", "first", "min", "max", "mean")
)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{biomarker_field}{Template field (i.e. the column name) used to look up biomarker values, e.g. \code{f.30750.0.0.Glycated_haemoglobin_HbA1c}}

\item{combine_instances}{The method used by \code{\link[=instance_combiner]{instance_combiner()}} or other Reduce-like functions when combining results of a lookup (e.g. a medication lookup, icd10 lookup, biomarker lookup) across multiple instances.

For example, when looking up whether a participant is on a medication, the result may differ depending on the instance number. In such a case, one would want to apply the \code{"any"} method so that results for instance 2 will be \code{\link[=Reduce]{Reduce()}}-ed with the \code{or} operator applied to the results of instances 0 and 1. In the case of numeric lookups (e.g. biomarkers that are recorded at multiple instances), one might want to use the \code{"mean"} method to average results across instances.

Can be one of:
\itemize{
\item \code{"any"} - use Boolean \code{or} (note: requires that lookup results are \code{logical})
\item \code{"min"} - use the minimum non-NA value
\item \code{"max"} - use the maximum non-NA value
\item \code{"first"} - use the first/earliest non-NA value
\item \code{"last"} - use the last/latest non-NA value
\item \code{"mean"} - use the mean of non-NA values
}

Functions that call \code{combine_instances()} may restrict the choice of options (e.g. it doesn't make sense to apply \code{"any"} to numeric data).}

\item{up_to_instance}{An integer or string specifying the instance number, or the <\code{\link[=dplyr_data_masking]{data-masking}}> name of the column containing instance numbers to use as the maximum allowed instance. Same syntax as \code{instance_num}, but used to include all data \emph{up to and including} the specified instance number.

The instance number is also known as the assessment visit number, of which there are 4 (labeled 0...3) in the UK BioBank study. This column does not exist in a standard distribution of the UKB data set, and should be defined and joined to the data set before performing analyses with this function.}

\item{...}{Arguments passed from specific \code{biomarker_} lookup functions to \code{\link[=biomarker_lookup]{biomarker_lookup()}}.}
}
\description{
Biomarker lookups
}
