% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Death_status.R
\name{Death_status}
\alias{Death_status}
\title{Death status}
\usage{
Death_status(data, .death_date_field = f.40000.0.0.Date_of_death)
}
\arguments{
\item{data}{input data}

\item{.death_date_field}{death field}
}
\description{
Death status
}
