% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_lookup.R
\name{diagnosis_lookup}
\alias{diagnosis_lookup}
\title{Diagnosis (self-reported) Lookup}
\usage{
diagnosis_lookup(
  data,
  ...,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported
)
}
\arguments{
\item{data}{data table}

\item{...}{Diagnosis codes to look up}

\item{.icd10_field}{example diagnosis field}
}
\description{
Diagnosis (self-reported) Lookup
}
