% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicity.R
\name{ethnicity_self_reported}
\alias{ethnicity_self_reported}
\title{Self-reported ethnicity, simplified}
\usage{
ethnicity_self_reported(data, ethnic_bkg = f.21000.0.0.Ethnic_background)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{ethnic_bkg}{initial ethnic background column}
}
\description{
Self-reported ethnicity, simplified
}
