% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_array_helper.R
\name{get_combiner_fn}
\alias{get_combiner_fn}
\title{Get Combiner Function}
\usage{
get_combiner_fn(
  combine_instances = c("any", "max", "min", "first", "last", "mean")
)
}
\arguments{
\item{combine_instances}{The method used by \code{\link[=instance_combiner]{instance_combiner()}} or other Reduce-like functions when combining results of a lookup (e.g. a medication lookup, icd10 lookup, biomarker lookup) across multiple instances.

For example, when looking up whether a participant is on a medication, the result may differ depending on the instance number. In such a case, one would want to apply the \code{"any"} method so that results for instance 2 will be \code{\link[=Reduce]{Reduce()}}-ed with the \code{or} operator applied to the results of instances 0 and 1. In the case of numeric lookups (e.g. biomarkers that are recorded at multiple instances), one might want to use the \code{"mean"} method to average results across instances.

Can be one of:
\itemize{
\item \code{"any"} - use Boolean \code{or} (note: requires that lookup results are \code{logical})
\item \code{"min"} - use the minimum non-NA value
\item \code{"max"} - use the maximum non-NA value
\item \code{"first"} - use the first/earliest non-NA value
\item \code{"last"} - use the last/latest non-NA value
\item \code{"mean"} - use the mean of non-NA values
}

Functions that call \code{combine_instances()} may restrict the choice of options (e.g. it doesn't make sense to apply \code{"any"} to numeric data).}
}
\description{
Helper to define a reduction function for combining instances and arrays. Returns a binary function based on the specified options. Typically used by \code{\link[=instance_combiner]{instance_combiner()}} and \code{\link[=reduce_by_row]{reduce_by_row()}}.
}
