% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICD10_lookup.R
\name{ICD10_lookup}
\alias{ICD10_lookup}
\alias{ICD10_heart_failure}
\alias{ICD10_pulm_htn}
\alias{ICD10_pulm_embolism}
\alias{ICD10_pulm_misc}
\alias{ICD10_pulm_ILD}
\alias{ICD10_pulm_emphysema}
\alias{ICD10_heart_pulm_transplant}
\title{ICD10 Lookup}
\usage{
ICD10_lookup(data, ..., .icd10_field = f.41270.0.0.Diagnoses_ICD10)

ICD10_heart_failure(data, .icd10_field = f.41270.0.0.Diagnoses_ICD10)

ICD10_pulm_htn(data, .icd10_field = f.41270.0.0.Diagnoses_ICD10)

ICD10_pulm_embolism(data, .icd10_field = f.41270.0.0.Diagnoses_ICD10)

ICD10_pulm_misc(data, .icd10_field = f.41270.0.0.Diagnoses_ICD10)

ICD10_pulm_ILD(data, .icd10_field = f.41270.0.0.Diagnoses_ICD10)

ICD10_pulm_emphysema(data, .icd10_field = f.41270.0.0.Diagnoses_ICD10)

ICD10_heart_pulm_transplant(data, .icd10_field = f.41270.0.0.Diagnoses_ICD10)
}
\arguments{
\item{data}{data table}

\item{...}{ICD10 codes to look up}

\item{.icd10_field}{example ICD10 field}
}
\description{
ICD10 Lookup
}
