% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_array_helper.R
\name{instance_combiner}
\alias{instance_combiner}
\title{Helper to apply a function up to a certain instance}
\usage{
instance_combiner(
  data,
  lookup_by_instance_fn,
  combine_instances = c("any", "max", "min", "first", "last", "mean"),
  up_to_instance = 3,
  after_instance = 0
)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{lookup_by_instance_fn}{Function that takes a target instance as its only argument, and returns a vector of data.}

\item{combine_instances}{The method used by \code{\link[=instance_combiner]{instance_combiner()}} or other Reduce-like functions when combining results of a lookup (e.g. a medication lookup, icd10 lookup, biomarker lookup) across multiple instances.

For example, when looking up whether a participant is on a medication, the result may differ depending on the instance number. In such a case, one would want to apply the \code{"any"} method so that results for instance 2 will be \code{\link[=Reduce]{Reduce()}}-ed with the \code{or} operator applied to the results of instances 0 and 1. In the case of numeric lookups (e.g. biomarkers that are recorded at multiple instances), one might want to use the \code{"mean"} method to average results across instances.

Can be one of:
\itemize{
\item \code{"any"} - use Boolean \code{or} (note: requires that lookup results are \code{logical})
\item \code{"min"} - use the minimum non-NA value
\item \code{"max"} - use the maximum non-NA value
\item \code{"first"} - use the first/earliest non-NA value
\item \code{"last"} - use the last/latest non-NA value
\item \code{"mean"} - use the mean of non-NA values
}

Functions that call \code{combine_instances()} may restrict the choice of options (e.g. it doesn't make sense to apply \code{"any"} to numeric data).}

\item{up_to_instance}{An integer or string specifying the instance number, or the <\code{\link[=dplyr_data_masking]{data-masking}}> name of the column containing instance numbers to use as the maximum allowed instance. Same syntax as \code{instance_num}, but used to include all data \emph{up to and including} the specified instance number.

The instance number is also known as the assessment visit number, of which there are 4 (labeled 0...3) in the UK BioBank study. This column does not exist in a standard distribution of the UKB data set, and should be defined and joined to the data set before performing analyses with this function.}
}
\description{
Helper to apply a function up to a certain instance
}
