% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{measurement_lookup}
\alias{measurement_lookup}
\alias{physio_systolicBP}
\alias{physio_diastolicBP}
\alias{physio_height_cm}
\alias{physio_weight_kg}
\title{General measurement lookups}
\usage{
measurement_lookup(
  data,
  measurement_field,
  combine_instances = c("last", "first", "min", "max", "mean"),
  combine_array = c("last", "first", "min", "max", "mean"),
  up_to_instance = 3
)

physio_systolicBP(
  data,
  measurement_field = f.4080.0.0.Systolic_blood_pressure_automated_reading,
  combine_array = "mean",
  ...
)

physio_diastolicBP(
  data,
  measurement_field = f.4079.0.0.Diastolic_blood_pressure_automated_reading,
  combine_array = "mean",
  ...
)

physio_height_cm(data, measurement_field = f.50.0.0.Standing_height, ...)

physio_weight_kg(data, measurement_field = f.21002.0.0.Weight, ...)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{measurement_field}{Template field (i.e. the column name) used to look up biomarker values, e.g. \code{f.4080.0.0.Systolic_blood_pressure_automated_reading}}

\item{combine_instances}{The method used by \code{\link[=up_to_instance_combiner]{up_to_instance_combiner()}} or other Reduce-like functions when combining results of a lookup (e.g. a medication lookup, icd10 lookup, biomarker lookup) across multiple instances.

For example, when looking up whether a participant is on a medication, the result may differ depending on the instance number. In such a case, one would want to apply the \code{"any"} method so that results for instance 2 will be \code{\link[=Reduce]{Reduce()}}-ed with the \code{or} operator applied to the results of instances 0 and 1. In the case of numeric lookups (e.g. biomarkers that are recorded at multiple instances), one might want to use the \code{"mean"} method to average results across instances.

Can be one of:
\itemize{
\item \code{"any"} - use Boolean \code{or} (note: requires that lookup results are \code{logical})
\item \code{"min"} - use the minimum non-NA value
\item \code{"max"} - use the maximum non-NA value
\item \code{"first"} - use the first/earliest non-NA value
\item \code{"last"} - use the last/latest non-NA value
\item \code{"mean"} - use the mean of non-NA values
}

Functions that call \code{combine_instances()} may restrict the choice of options (e.g. it doesn't make sense to apply \code{"any"} to numeric data).}

\item{combine_array}{If a measurement field has multiple array values (e.g. blood pressure recordings are made in duplicate at each instance), specify how these values should be combined. See \code{combine_instances} for details.}

\item{up_to_instance}{An integer or string specifying the instance number, or the <\code{\link[=dplyr_data_masking]{data-masking}}> name of the column containing instance numbers to use as the maximum allowed instance. Same syntax as \code{instance_num}, but used to include all data \emph{up to and including} the specified instance number.

The instance number is also known as the assessment visit number, of which there are 4 (labeled 0...3) in the UK BioBank study. This column does not exist in a standard distribution of the UKB data set, and should be defined and joined to the data set before performing analyses with this function.}

\item{...}{Arguments passed from specific \code{physio_measurement_} lookup functions to \code{\link[=measurement_lookup]{measurement_lookup()}}.}
}
\description{
General measurement lookups
}
