% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medications.R
\name{medication_lookup}
\alias{medication_lookup}
\title{Medication Lookups}
\usage{
medication_lookup(
  data,
  codes,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{codes}{Medication codes to look up}

\item{after_instance}{An integer or string specifying an instance number, or the <\code{\link[=dplyr_data_masking]{data-masking}}> name of the column containing instance numbers to use as the MINIMUM (NON-INCLUSIVE)  instance. Same syntax as \code{instance_num}, but used to include all data \emph{after and NOT including} the specified instance number.

The instance number is also known as the assessment visit number, of which there are 4 (labeled 0...3) in the UK Biobank study. This column does not exist in a standard distribution of the UKB data set, and should be defined and joined to the data set before performing analyses with this function.}

\item{up_to_instance}{An integer or string specifying an instance number, or the <\code{\link[=dplyr_data_masking]{data-masking}}> name of the column containing instance numbers to use as the maximum allowed instance. Same syntax as \code{instance_num}, but used to include all data \emph{up to and including} the specified instance number.

The instance number is also known as the assessment visit number, of which there are 4 (labeled 0...3) in the UK Biobank study. This column does not exist in a standard distribution of the UKB data set, and should be defined and joined to the data set before performing analyses with this function.}

\item{medication_field}{e.g. \code{f.20003.0.0.Treatment_medication_code}}
}
\description{
Medication Lookups
}
