% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medications.R
\name{medication_lookup_by_disease}
\alias{medication_lookup_by_disease}
\alias{medication_statin}
\alias{medication_antiHLD}
\alias{medication_PDE5i}
\alias{medication_BB}
\alias{medication_ACEi}
\alias{medication_ARB}
\alias{medication_CCB}
\alias{medication_loopdiuretic}
\alias{medication_thiazide}
\alias{medication_antiHTN}
\alias{medication_insulin}
\alias{medication_oralhypoglycemic}
\alias{medication_pulm_endothelin}
\title{Medication lookups for selected diseases

These functions are aliases for \code{\link[=medication_lookup]{medication_lookup()}}, using hard-coded \code{codes} that have been pre-selected for each disease/medication class.}
\usage{
medication_statin(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)

medication_antiHLD(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)

medication_PDE5i(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)

medication_BB(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)

medication_ACEi(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)

medication_ARB(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)

medication_CCB(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)

medication_loopdiuretic(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)

medication_thiazide(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)

medication_antiHTN(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)

medication_insulin(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)

medication_oralhypoglycemic(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)

medication_pulm_endothelin(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  medication_field = f.20003.0.0.Treatment_medication_code
)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{up_to_instance}{An integer or string specifying the instance number, or the <\code{\link[=dplyr_data_masking]{data-masking}}> name of the column containing instance numbers to use as the maximum allowed instance. Same syntax as \code{instance_num}, but used to include all data \emph{up to and including} the specified instance number.

The instance number is also known as the assessment visit number, of which there are 4 (labeled 0...3) in the UK BioBank study. This column does not exist in a standard distribution of the UKB data set, and should be defined and joined to the data set before performing analyses with this function.}

\item{...}{Arguments passed to \code{\link[=medication_lookup]{medication_lookup()}} (excluding the \code{codes} argument, which is pre-set by the \code{medication_} lookup functions).

Specifically:
\itemize{
\item \code{data}
\item \code{up_to_instance = DEFAULT_UP_TO_INST}
\item \code{medication_field = f.20003.0.0.Treatment_medication_code}
}}
}
\description{
Medication lookups for selected diseases

These functions are aliases for \code{\link[=medication_lookup]{medication_lookup()}}, using hard-coded \code{codes} that have been pre-selected for each disease/medication class.
}
