% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{ukbiobank}
\alias{ukbiobank}
\title{ukbiobank}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, sex).}

\item{instance_num}{An integer or string specifying the instance number, or the <\code{\link[=dplyr_data_masking]{data-masking}}> name of the column containing instance numbers to use in the current function.

The instance number is also known as the assessment visit number, of which there are 4 (labeled 0...3) in the UK Biobank study. This column does not exist in a standard distribution of the UKB data set, and should be joined to the data set before performing analyses with this function.}

\item{date}{Target date, as a string which is then passed to \code{lubridate}.}
}
\description{
his package contains a collection of tools and functions to facilitatoe analysis of UK Biobank phenotype data.
}
