% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{years_from_date_to_death}
\alias{years_from_date_to_death}
\title{Years from date to death}
\usage{
years_from_date_to_death(
  data,
  date,
  death_date_field = f.40000.0.0.Date_of_death
)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{date}{Target date, as a string which is then passed to \code{\link[lubridate:as_date]{lubridate::as_date()}}.}

\item{death_date_field}{f.40000.0.0.Date_of_death}
}
\description{
Years from date to death
}
