% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Death_status.R
\name{Years_from_instance_to_death}
\alias{Years_from_instance_to_death}
\title{Years from instance to death}
\usage{
Years_from_instance_to_death(
  data,
  instance_column,
  .instance_date_field = f.53.0.0.Date_of_attending_assessment_centre,
  .death_date_field = f.40000.0.0.Date_of_death
)
}
\arguments{
\item{data}{input data}

\item{instance_column}{instance column name}

\item{.instance_date_field}{e.g. f.53.0.0.Date_of_attending_assessment_centre}

\item{.death_date_field}{e.g. f.40000.0.0.Date_of_death}
}
\description{
Years from instance to death
}
