% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_at_date.R
\name{age_at_date}
\alias{age_at_date}
\alias{age_group_at_date}
\title{Age at date}
\usage{
age_at_date(
  data,
  date,
  .year_of_birth = f.34.0.0.Year_of_birth,
  .month_of_birth = f.52.0.0.Month_of_birth
)

age_group_at_date(
  data,
  date,
  .year_of_birth = f.34.0.0.Year_of_birth,
  .month_of_birth = f.52.0.0.Month_of_birth
)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, sex).}

\item{date}{Target date, as a string which is then passed to \code{lubridate}.}

\item{.year_of_birth}{<\code{\link[=dplyr_data_masking]{data-masking}}> Name of column containing birth year}

\item{.month_of_birth}{<\code{\link[=dplyr_data_masking]{data-masking}}> Name of column containing birth month}
}
\value{
Numberic vector of age in years (\code{age_at_date}), or character vector of age groups (\code{age_group_at_date})
}
\description{
Get age (in years) as of a particular date. The alternate function, \strong{\code{age_group_at_date()}}, returns a string describing the age group in decades (e.g. "20-29") for a 24-year-old patient.
}
