% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age_at_date}
\alias{age_at_date}
\title{Age at date}
\usage{
age_at_date(
  data,
  date,
  year_of_birth_field = f.34.0.0.Year_of_birth,
  month_of_birth_field = f.52.0.0.Month_of_birth
)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{date}{Target date, as a string which is then passed to \code{\link[lubridate:as_date]{lubridate::as_date()}}.}

\item{year_of_birth_field}{<\code{\link[=dplyr_data_masking]{data-masking}}> Name of column containing birth year}

\item{month_of_birth_field}{<\code{\link[=dplyr_data_masking]{data-masking}}> Name of column containing birth month}
}
\value{
Numeric vector of age in years (\code{age_at_date()}), or character vector of age groups (\code{age_group_at_date()})
}
\description{
Get age (in years) as of a particular date.
}
\details{
NOTE: This result is approximate, since only the year and month of birth are provided. The day of birth is assumed to be 15.
}
