% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Age_at_instance.R
\name{Age_at_instance}
\alias{Age_at_instance}
\alias{Age_group_at_instance}
\title{Age at instance}
\usage{
Age_at_instance(
  data,
  instance_column,
  .age_at_assessment_col = f.21003.0.0.Age_when_attended_assessment_centre
)

Age_group_at_instance(
  data,
  instance_column,
  .age_at_assessment_col = f.21003.0.0.Age_when_attended_assessment_centre
)
}
\arguments{
\item{data}{input data}

\item{instance_column}{name of column containing instance numbers}

\item{.age_at_assessment_col}{age at assessment}
}
\description{
Age at instance
}
