% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{convert_to_age_group}
\alias{convert_to_age_group}
\title{Age group from vector of ages}
\usage{
convert_to_age_group(age)
}
\arguments{
\item{age}{Input vector of ages}
}
\description{
Age group from vector of ages
}
