% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{death_status}
\alias{death_status}
\title{Death status (true/false)}
\usage{
death_status(data, death_date_col = f.40000.0.0.Date_of_death)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{death_date_col}{Template column name for date of death.\cr Default = \code{f.40000.0.0.Date_of_death}.}
}
\description{
Returns whether a death event has been reported for each participant.
}
\section{Column names}{
 Most functions in this package will take column names (\verb{*_col}) as optional arguments (otherwise a default column names are used) which are then used as templates to find all other columns with the same field number, but different instance and array numbers. These functions will automatically find all matching instances (and arrays within each instance) within the specified parameters. Internally, a set of \verb{expand_instance_*()} helper functions, which themselves rely on the \code{column_expansion_helper()} function, perform the work of searching for matching columns.
}

\seealso{
\code{\link[=years_from_date_to_death]{years_from_date_to_death()}}, \code{\link[=years_from_instance_to_death]{years_from_instance_to_death()}}
}
