% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{death_status}
\alias{death_status}
\title{Death status}
\usage{
death_status(data, death_date_field = f.40000.0.0.Date_of_death)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{death_date_field}{death field}
}
\description{
Death status
}
