% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_lookup.R
\name{diagnosis_lookup}
\alias{diagnosis_lookup}
\alias{diagnosed_chf}
\alias{diagnosed_htn}
\alias{diagnosed_hld}
\alias{diagnosed_copd}
\title{Diagnosis (self-reported) Lookup}
\usage{
diagnosis_lookup(
  data,
  ...,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported
)

diagnosed_chf(
  data,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported,
  .icd10_field = f.41270.0.0.Diagnoses_ICD10
)

diagnosed_htn(
  data,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported,
  .icd10_field = f.41270.0.0.Diagnoses_ICD10
)

diagnosed_hld(
  data,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported,
  .icd10_field = f.41270.0.0.Diagnoses_ICD10
)

diagnosed_copd(
  data,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported,
  .icd10_field = f.41270.0.0.Diagnoses_ICD10
)
}
\arguments{
\item{data}{data table}

\item{...}{Diagnosis codes to look up}

\item{.icd10_field}{example diagnosis field}
}
\description{
Diagnosis (self-reported) Lookup
}
