% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_lookup.R
\name{diagnosis_lookup}
\alias{diagnosis_lookup}
\alias{diagnosed_chf}
\alias{diagnosed_htn}
\alias{diagnosed_hld}
\alias{diagnosed_copd}
\alias{diagnosed_pHTN}
\alias{diagnosed_pulm_embolism}
\alias{diagnosed_diabetes}
\title{Diagnosis (self-reported) Lookup}
\usage{
diagnosis_lookup(
  data,
  ...,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported
)

diagnosed_chf(
  data,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported,
  .icd10_field = f.41270.0.0.Diagnoses_ICD10
)

diagnosed_htn(
  data,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported,
  .icd10_field = f.41270.0.0.Diagnoses_ICD10
)

diagnosed_hld(
  data,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported,
  .icd10_field = f.41270.0.0.Diagnoses_ICD10
)

diagnosed_copd(
  data,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported,
  .icd10_field = f.41270.0.0.Diagnoses_ICD10
)

diagnosed_pHTN(
  data,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported,
  .icd10_field = f.41270.0.0.Diagnoses_ICD10
)

diagnosed_pulm_embolism(
  data,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported,
  .icd10_field = f.41270.0.0.Diagnoses_ICD10
)

diagnosed_diabetes(
  data,
  .diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported,
  .icd10_field = f.41270.0.0.Diagnoses_ICD10
)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, sex).}

\item{...}{Diagnosis codes to look up}

\item{.diagnosis_field}{e.g. f.20002.0.0.Non_cancer_illness_code_self_reported}

\item{.icd10_field}{e.g. f.41270.0.0.Diagnoses_ICD10}
}
\description{
Diagnosis (self-reported) Lookup
}
