% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoses.R
\name{dx_self_report_lookup}
\alias{dx_self_report_lookup}
\alias{diagnosed_chf}
\alias{diagnosed_htn}
\alias{diagnosed_hld}
\alias{diagnosed_copd}
\alias{diagnosed_pHTN}
\alias{diagnosed_pulm_embolism}
\alias{diagnosed_diabetes}
\title{Diagnosis lookup}
\usage{
dx_self_report_lookup(
  data,
  codes,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported
)

diagnosed_chf(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported
)

diagnosed_htn(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported,
  min_sbp = 140,
  min_dbp = 90
)

diagnosed_hld(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported
)

diagnosed_copd(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported
)

diagnosed_pHTN(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported
)

diagnosed_pulm_embolism(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported
)

diagnosed_diabetes(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  diagnosis_field = f.20002.0.0.Non_cancer_illness_code_self_reported
)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{codes}{Diagnosis codes to look up}

\item{up_to_instance}{An integer or string specifying the instance number, or the <\code{\link[=dplyr_data_masking]{data-masking}}> name of the column containing instance numbers to use as the maximum allowed instance. Same syntax as \code{instance_num}, but used to include all data \emph{up to and including} the specified instance number.

The instance number is also known as the assessment visit number, of which there are 4 (labeled 0...3) in the UK BioBank study. This column does not exist in a standard distribution of the UKB data set, and should be defined and joined to the data set before performing analyses with this function.}

\item{diagnosis_field}{e.g. f.20002.0.0.Non_cancer_illness_code_self_reported}
}
\description{
Diagnosis lookup
}
