% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ethnicity_self_reported.R
\name{Ethnicity_self_reported}
\alias{Ethnicity_self_reported}
\title{Self-reported ethnicity, simplified}
\usage{
Ethnicity_self_reported(data, .ethnic_bkg = f.21000.0.0.Ethnic_background)
}
\arguments{
\item{data}{input data}

\item{.ethnic_bkg}{initial ethnic background column}
}
\description{
Self-reported ethnicity, simplified
}
