% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{icd10_lookup}
\alias{icd10_lookup}
\alias{icd10_htn}
\alias{icd10_hld}
\alias{icd10_heart_failure}
\alias{icd10_pulm_htn}
\alias{icd10_pulm_embolism}
\alias{icd10_pulm_misc}
\alias{icd10_pulm_ILD}
\alias{icd10_copd}
\alias{icd10_heart_pulm_transplant}
\alias{icd10_diabetes}
\title{ICD10 Lookup}
\usage{
icd10_lookup(
  data,
  codes,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  icd10_field = f.41270.0.0.Diagnoses_ICD10,
  icd10_date_field = f.41280.0.0.Date_of_first_in_patient_diagnosis_ICD10,
  date_of_instance_field = f.53.0.0.Date_of_attending_assessment_centre
)

icd10_htn(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  icd10_field = f.41270.0.0.Diagnoses_ICD10
)

icd10_hld(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  icd10_field = f.41270.0.0.Diagnoses_ICD10
)

icd10_heart_failure(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  icd10_field = f.41270.0.0.Diagnoses_ICD10
)

icd10_pulm_htn(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  icd10_field = f.41270.0.0.Diagnoses_ICD10
)

icd10_pulm_embolism(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  icd10_field = f.41270.0.0.Diagnoses_ICD10
)

icd10_pulm_misc(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  icd10_field = f.41270.0.0.Diagnoses_ICD10
)

icd10_pulm_ILD(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  icd10_field = f.41270.0.0.Diagnoses_ICD10
)

icd10_copd(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  icd10_field = f.41270.0.0.Diagnoses_ICD10
)

icd10_heart_pulm_transplant(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  icd10_field = f.41270.0.0.Diagnoses_ICD10
)

icd10_diabetes(
  data,
  after_instance = DEFAULT_AFTER_INST,
  up_to_instance = DEFAULT_UP_TO_INST,
  icd10_field = f.41270.0.0.Diagnoses_ICD10
)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{codes}{ICD10 codes to look up}

\item{after_instance}{An integer or string specifying an instance number, or the <\code{\link[=dplyr_data_masking]{data-masking}}> name of the column containing instance numbers to use as the MINIMUM (NON-INCLUSIVE)  instance. Same syntax as \code{instance_num}, but used to include all data \emph{after and NOT including} the specified instance number.

The instance number is also known as the assessment visit number, of which there are 4 (labeled 0...3) in the UK Biobank study. This column does not exist in a standard distribution of the UKB data set, and should be defined and joined to the data set before performing analyses with this function.}

\item{up_to_instance}{An integer or string specifying an instance number, or the <\code{\link[=dplyr_data_masking]{data-masking}}> name of the column containing instance numbers to use as the maximum allowed instance. Same syntax as \code{instance_num}, but used to include all data \emph{up to and including} the specified instance number.

The instance number is also known as the assessment visit number, of which there are 4 (labeled 0...3) in the UK Biobank study. This column does not exist in a standard distribution of the UKB data set, and should be defined and joined to the data set before performing analyses with this function.}

\item{icd10_field}{example ICD10 field}

\item{icd10_date_field}{example ICD10 date field}

\item{date_of_instance_field}{Template field (using <\code{\link[=dplyr_data_masking]{data-masking}}> rules) for the date associated with each instance.}
}
\description{
ICD10 Lookup
}
