% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medications.R
\name{medication_lookup_by_disease}
\alias{medication_lookup_by_disease}
\alias{medication_statin}
\alias{medication_antiHLD}
\alias{medication_PDE5i}
\alias{medication_BB}
\alias{medication_ACEi}
\alias{medication_ARB}
\alias{medication_CCB}
\alias{medication_loopdiuretic}
\alias{medication_thiazide}
\alias{medication_antiHTN}
\alias{medication_insulin}
\alias{medication_oralhypoglycemic}
\alias{medication_pulm_endothelin}
\title{Medication lookups for selected diseases

These functions are aliases for \code{\link[=medication_lookup]{medication_lookup()}}, using hard-coded \code{codes} that have been pre-selected for each disease/medication class.}
\usage{
medication_statin(data, ...)

medication_antiHLD(data, ...)

medication_PDE5i(data, ...)

medication_BB(data, ...)

medication_ACEi(data, ...)

medication_ARB(data, ...)

medication_CCB(data, ...)

medication_loopdiuretic(data, ...)

medication_thiazide(data, ...)

medication_antiHTN(data, ...)

medication_insulin(data, ...)

medication_oralhypoglycemic(data, ...)

medication_pulm_endothelin(data, ...)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{...}{Arguments passed to \code{\link[=medication_lookup]{medication_lookup()}} (excluding the \code{codes} argument, which is pre-set by the \code{medication_} lookup functions).

Specifically:
\itemize{
\item \code{data}
\item \code{up_to_instance = 3}
\item \code{medication_field = f.20003.0.0.Treatment_medication_code}
}}
}
\description{
Medication lookups for selected diseases

These functions are aliases for \code{\link[=medication_lookup]{medication_lookup()}}, using hard-coded \code{codes} that have been pre-selected for each disease/medication class.
}
