% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Death_status.R
\name{Years_from_date_to_death}
\alias{Years_from_date_to_death}
\title{Years from date to death}
\usage{
Years_from_date_to_death(
  data,
  date,
  .death_date_field = f.40000.0.0.Date_of_death
)
}
\arguments{
\item{data}{Input data}

\item{date}{Start date}

\item{.death_date_field}{f.40000.0.0.Date_of_death}
}
\description{
Years from date to death
}
