% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{years_from_instance_to_death}
\alias{years_from_instance_to_death}
\title{Years from instance to death}
\usage{
years_from_instance_to_death(
  data,
  instance_num,
  instance_date_field = f.53.0.0.Date_of_attending_assessment_centre,
  death_date_field = f.40000.0.0.Date_of_death
)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{instance_num}{instance column name}

\item{instance_date_field}{e.g. f.53.0.0.Date_of_attending_assessment_centre}

\item{death_date_field}{e.g. f.40000.0.0.Date_of_death}
}
\description{
Years from instance to death
}
