% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{years_from_instance_to_death}
\alias{years_from_instance_to_death}
\title{Years from instance to death}
\usage{
years_from_instance_to_death(
  data,
  instance_num,
  date_of_instance_field = f.53.0.0.Date_of_attending_assessment_centre,
  death_date_field = f.40000.0.0.Date_of_death
)
}
\arguments{
\item{data}{The primary data frame.

This data frame includes all necessary columns required to perform look-ups (e.g. ICD10 code columns, medication code columns, age, sex, etc.).}

\item{instance_num}{An integer or string specifying the instance number, or the <\code{\link[=dplyr_data_masking]{data-masking}}> name of the column containing instance numbers to use in the current function.

The instance number is also known as the assessment visit number, of which there are 4 (labeled 0...3) in the UK Biobank study. This column does not exist in a standard distribution of the UKB data set, and should be defined and joined to the data set before performing analyses with this function.}

\item{date_of_instance_field}{Template field (using <\code{\link[=dplyr_data_masking]{data-masking}}> rules) for the date associated with each instance.}

\item{death_date_field}{e.g. f.40000.0.0.Date_of_death}
}
\description{
Years from instance to death
}
